gernated files with _pmr suffix
